モデル追加したり、サウンドエンジン追加したり

+++ GameTemplate/Game/Collection.h
@@ -26,7 +26,7 @@ private:
 	ID3D11SamplerState* m_samplerState = nullptr;		//サンプラステート。
 	PostEffect m_postEffect[W_NUMBER * H_NUMBER];
 	ID3D11RasterizerState* m_rasterizerState = NULL;	//ラスタライザステート。
-	std::map<int, SkinModel_ObjData*> m_modelList;
+	std::map<float, SkinModel_ObjData*> m_modelList;
 	float m_degree = 0.0f;
 	int m_page = 1;
 	int m_listSize = 0;
